added driver signup, driver add and more
