+++ delivery_app/lib/main.dart
@@ -1,4 +1,4 @@
-import 'package:delivery_app/merchant_order.dart';
+import 'package:delivery_app/merchant_page.dart';
 import 'package:delivery_app/test.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
@@ -10,7 +10,7 @@ import 'package:delivery_app/welcome.dart';
 
 
 void main() => runApp(MaterialApp(
-    home: MerchantOrder(),debugShowCheckedModeBanner: false,
+    home: MyApp(),debugShowCheckedModeBanner: false,
 ));
 //SystemChrome.setEnabledSystemUIOverlays([]);
